gaze-detection/requirements.txt
fastapi
torch
torchvision
uvicorn
pillow
kafka-python
